2-dsc_streamlit/streamlit/requirements.txt
streamlit
numpy
opencv-python
pillow
python-multipart
uvicorn
